Unreviewed, fix -Wformat warning https://bugs.webkit.org/show_bug.cgi?id=195895 <rdar://problem/48517629>
* bmalloc/Scavenger.cpp:
(bmalloc::Scavenger::threadRunLoop):

Canonical link: https://commits.webkit.org/210234@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243165 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/bmalloc/bmalloc/Scavenger.cpp
@@ -356,7 +356,7 @@ void Scavenger::threadRunLoop()
         }
 
         if (verbose)
-            fprintf(stderr, "new wait time %lldms\n", m_waitTime.count());
+            fprintf(stderr, "new wait time %lldms\n", static_cast<long long int>(m_waitTime.count()));
     }
 }
